buckets: implement "max-keys" when listing buckets
If max-keys is not an integer, we raise and `InvalidArgument`
error. Otherwise, it is used to limit the number of keys returned. While
the documentation is a bit unclear, the number of keys returned seems to
encompass both the common prefixes and the content entries.

To stay compatible with the previous version, there is no hard limit by
default. Therefore, there is also no limit to this value. If not value
is provided, we return the number of provided elements rounded to the
next hundred. The `IsTruncated` flag is also handled.

All this is done late in the processing, mostly because we can set the
`IsTruncated` easily at this stage. In the earlier stages, only lazy
sequences are used so it should be performant enough.