http: do not use white spaces to keep the connection alive
When sending either `\n` or ` ` to keep the HTTP connection alive, the
answer is also padded with those characters. An XML answer cannot start
with a whitespace. If there is an XML declaration, it must be first.

Boto is using Python XML SAX parser which doesn't like whitespaces. XML
parser in PHP doesn't like it either.

Unfortunately, with this change, Jetty doesn't send an empty chunk, it
doesn't send anything. So, this breaks the keepalived mechanism.