sig: fix signature when both `Date` and `x-amz-date` headers are present
When both headers are present, the `Date` header should be ignored and
the corresponding line in the signature should be empty. This
modification makes the appropriate test succeed.

We assume that when the `expires` query string is present, it overrides
the `Date` field even when `x-amz-date` is present. This is a corner
case not correctly specified in Amazon S3 documentation since most of
the time, the `x-amz-date` header won't be present.