meta: include object itself in prefix search. fixes #117 (#125)
In object lists, when the given prefix is itself a key, it
is excluded from search results. This introduces a behavioral change
from AWS S3's approach and breaks s3cmd for cases such as:

   s3cmd ls s3://bucket/object.txt

Provided here is a slight modification to the fetcher callback
signature, accounting for an `init?` argument where queries
may include the prefix itself in searches.